ci: Don't create GH releases for main builds
Use the `disable` field and some nasty go templating to prevent goreleaser from
creating GitHub releases when we run the release workflow on the main branch
(and release branches). We only want GH releases (and the associated tags) when
the workflow runs on a tag we've created for release.

While we're here, set `prerelease` to auto, which should automatically mark
releases as prereleases if they have a release semver. I thought this was the
default, but it's not.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>