ci: Don't fail fast in the AWS e2e tests
Our AWS-dependent e2e tests hit transient failures relatively often, due to some
combination of runners/Crossplane being slow and AWS taking a while to reconcile
things. Currently, if one of these jobs fails any other in-progress or pending
ones from the matrix get canceled, meaning we have to re-run all of them if it's
a transient failure.

Set `fail-fast: false` so that the rest of the matrix can run even when some
jobs fail. This should reduce the number of jobs we have to re-run due to
transient failures.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>