ci: Generate docs and create a docs PR as part of release
As part of doing a release, check out the docs repo, run our docs generator, and
create a draft PR with the updates. We still have to manually mark the PR ready
for review (once we promote the release to stable) and merge it, but it's still
nice to get rid of a manual step.

Fixes PEX-44

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>