projects: Generate meta schemas in `up project build`
We used to generate schemas for our Upbound metadata APIs (project and test
types) as part of `up project build` as well as in the `up test` commands. We
still do it in `up test`, but lost it in `up project build` as part of the v0.40
schema generation changes.

Re-introduce meta API schema generation to `up project build` so that we keep
the behavior from before.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>