release: Don't assume `git describe` is the version
It used to be the case that `git describe` was always the version we were
releasing. Now we generate temporary version tags for main and release branch
builds, and these tags don't get pushed back to GitHub, so the promote job
doesn't know what they were.

Use goreleaser's metadata output to make the version an output from the build
job and consume it in the promote job. This guarantees that we're always
promoting the version we just built.

Also update the write-version script to use goreleaser's version string. This
was already working correctly, but it feels safer to not rely on `git describe`.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>