test: Support go-templating in `up test generate` and `up test run`
Let users write tests as plain YAML with go-templating. We don't provide any
inputs to the templating, but users can do loops and conditionals, define
variables, etc.

Generate a special jsonschema for tests so users get good completion for our
test types when writing the items array. Use this as the schema for test files
instead of the `index.schema.json` we use for functions, since it contains all
types and does not know about the special "items".

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>