ci: Make release work properly
In an ideal world, we'd kick off a release by creating a tag via the GitHub UI
and let the release workflow run on the newly created tag. But, there's no way
to create a tag without creating a GH release, and we want to let goreleaser do
that for us, so we have to create a tag via actions.

Previously, we had a separate tag action, which invoked the release action via a
`workflow_run` trigger after it ran. This didn't work properly, since the ref
for a `workflow_run` triggered action is the default branch, not the
newly-created tag or even the branch the previous workflow ran on.

We already have a step in the release workflow to create a (local) tag, since we
need it for main branch builds. Get rid of the tag action and add the version to
release as an input to the release action via `workflow_dispatch`. Create the
verison tag locally if it's given, and then let goreleaser create the release
and tag in GH when it runs.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>