projects: Unify supported languages and test languages
We now have parity between languages we support for functions and for tests, and
we should make sure this doesn't diverge again in the future. To that end, use
the same list of supported languages for both purposes in `up project init` and
remove the "Supported for tests" column in the docs.

While we're here, add logic to the wizard to set the default test language
option to the language the user selected for functions whenever possible.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>