project: Add support for extra resources in `up project run`
Make it possible to create extra resources (for example, pull secrets and
ImageConfigs) as part of `up project run` by introducing a new
`--extra-resources` flag. The flag can be specified more than once to provide an
arbitrary number of manifests or directories of manifests. Extra resources are
applied to the dev control plane before the project is installed.

Fixes CTP-52

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>