login: Simplify callback listener creation
The latest golangci-lint flagged an issue with our use of `net.Listen`, which
made me re-evaluate this code overall. It had a race condition in that we let
the system find a free port, then gave that port up, then tried to use it
again. It was also constructing an unnecessary server mux and doing an
unnecessary type conversion on the port.

Appease the linter by using `net.ListenConfig.Listen` (which takes a context),
and then actually use the returned listener instead of closing it right
away. Use the callback handler as the server's handler instead of wrapping it
with a mux. Keep the port as a string instead of converting to an int since
we're only going to print it anyway.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>