*: Replace pterm interactive elements with huh
We recently hit a [pterm text input bug] that (a) looks glitchy and (b) results
in us getting corrupted input from the user. The pterm library is poorly
maintained and uses a ton of non-obvious global variables, among other issues
that make it not concurrency safe. These bugs are hard to find with automated
tests, since pterm is largely used for user interaction.

Replace all the interactive elements we use from pterm with the charmbracelet
huh package. We've found charmbracelet libraries to be stable in general, and
they are actively developed and maintained. This slightly changes the look and
feel of our prompts etc., but shouldn't be too jarring a change for users.

Fixes CTP-98

References:
[pterm text input bug]: https://github.com/pterm/pterm/issues/750

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>