Replace all pterm usage with upterm
Introduce a new `upterm.Printer` type that does all the printing things we
need. Initialize it once in the top-level `up` command, respecting the
`--pretty` and `--quiet` flags, then bind it to kong so that all commands can
use it. Stop using the pterm library everywhere in favor of the new
`upterm.Printer`. This makes us more consistent in respecting the output-related
flags (though still not perfect) and gets rid of the problematic pterm library.

Leave `up space connect` and `up space disconnect` as-is, since they do crazy
stuff with pterm that we don't need/want to support in `upterm`. We can rework
these someday, but given how hard they are to test I don't want to do it now.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>