ctp: Correctly create a group-scoped client when in a control plane context
When the current context is a control plane, commands like `up ctp list` should
use the control plane's group. Previously, we weren't correctly constructing a
group-scoped client, so we ended up using the deafult group. This happened
because the kubeconfig's namespace gets lost when it's converted to a
restconfig.

Explicitly create a namespaced client when we're in a group context, so that we
use the right group for these commands.

Fixes CTP-106

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>